# Prefer String#tr for other crockford features