--- a/benchmark/compare_with_othergems/rafaelsales/parser.rb
+++ b/benchmark/compare_with_othergems/rafaelsales/parser.rb
@@ -1,0 +1,9 @@
+# coding: utf-8
+# frozen_string_literal: true
+
+require('benchmark/ips')
+require('ulid')
+
+raise("Bug to setup: #{ULID.methods(false)}") unless ULID.const_defined?(:Generator)
+
+p("`ulid gem - #{ULID::VERSION}` does not have parsers yet")