--- a/lib/ulid/errors.rb
+++ b/lib/ulid/errors.rb
@@ -1,0 +1,10 @@
+# coding: us-ascii
+# frozen_string_literal: true
+# shareable_constant_value: literal
+
+class ULID
+  class Error < StandardError; end
+  class OverflowError < Error; end
+  class ParserError < Error; end
+  class UnexpectedError < Error; end
+end