--- a/sig/ulid.rbs
+++ b/sig/ulid.rbs
@@ -1,6 +1,5 @@
 class ULID < Object
   VERSION: String
-  CROCKFORD_BASE32_ENCODING_STRING: String
   TIMESTAMP_ENCODED_LENGTH: 10
   RANDOMNESS_ENCODED_LENGTH: 16
   ENCODED_LENGTH: 26
@@ -39,13 +38,16 @@ class ULID < Object
     class SetupError < UnexpectedError
     end
 
+    ENCODING_STRING: String
+    ORDERED_CROCKFORD_BASE32_CHARS: Array[String]
+    ORDERED_N32_CHARS: Array[String]
     N32_CHAR_BY_CROCKFORD_BASE32_CHAR: Hash[String, String]
     CROCKFORD_BASE32_CHAR_PATTERN: Regexp
     CROCKFORD_BASE32_CHAR_BY_N32_CHAR: Hash[String, String]
-    ORDERED_CROCKFORD_BASE32_CHARS: String
-    ORDERED_N32_CHARS: String
-    STANDARD_BY_VARIANT: Hash[String, String]
-    VARIANT_PATTERN: Regexp
+    CROCKFORD_BASE32_TR_PATTERN: String
+    N32_TR_PATTERN: String
+    VARIANT_TR_PATTERN: String
+    NORMALIZED_TR_PATTERN: String
 
     # A private API. Should not be used in your code.
     def self.encode: (Integer integer) -> String